Collider's gap is now made of floats

+++ src/include/FROG/Collision/RoundCollider.hpp
@@ -17,12 +17,12 @@ namespace frog{
     /// the bounding box
     sf::Vector2f center;
     /// the gap between origin of parent and origin of the box
-    sf::Vector2i gap;
+    sf::Vector2f gap;
 
 
   public:
     RoundCollider(const sf::Vector2f& rad = sf::Vector2f(1.0f, 1.0f),
-                  const sf::Vector2i& gap = sf::Vector2i(0,0),
+                  const sf::Vector2f& gap = sf::Vector2f(0,0),
                   std::function<void(Collision)> = [](Collision){} );
     virtual ~RoundCollider();
     virtual sf::FloatRect getBoundingBox() const;
@@ -32,7 +32,7 @@ namespace frog{
     virtual float getYMax() const;
     virtual void update(const ComponentHolder&);
     virtual void resize(const sf::Vector2f&);
-    virtual void setGap(const sf::Vector2i&);
+    virtual void setGap(const sf::Vector2f&);
 
   };
